requirements.txt
aiohttp>=3.8
beautifulsoup4>=4.12
python-telegram-bot==20.1
pyyaml>=6.0
